Adding a parity generator
At this stage, the parity generator is not fully functional. It can tell you the number of bits but needs work to display whether the number was negative or positive.  Additionally, there needs to be some sort of code to decode the parity generator. Though this should be simple as the code can just be done in reverse. 
EH